ovirt: perform snapshot removal sequentially
Because of ovirt-engine's internal locks, starting a snapshot removal
operation will lock the VM, and will prevent other snapshot removals on
the same VM, leaving all snapshots around except for the first one.

This patch waits for each snapshot removal to complete before starting
the next one. And if the first one fails it would retry 3 times, because
there is a possible race with image transfer, as CDI does not currently
wait correctly for the transfer to finish.

Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>