virt-v2v: Let virt-v2v do the copying and conversion
Modify container entrypoint to let virt-v2v do the data copying as well
as conversion instead of the sort of in-place conversion we used to do
before.

To achieve that we use the `-o local` output mode. For that output mode
to work as we need we have to prepare some symlinks for disks. Our disks
are expected to be either in `/mnt/disks/disk*/disk.img` files on
filesystem PVs or on `/dev/block*` devices for block based PVs. On the
other hand virt-v2v expects the disks to be in `<vm_name>-sd<X>` files.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>