Migrate forklift-operator to forklift with bazel
This patch changes the build processes to the bazel tool, because of it
we need to change behaviour of the builds.

The step by step description of builds:

Operator build:
- Build the oprator image with ansible dependencies and with roles.

Bundle build:
- Change to the operator dir
- Get the current date with which the operator will be built
- Build the config files with kustomize
- Substitute env variables in the generated config
- Generate the bundle from the config
- Build the image with the bundle

Index build:
- Substitute env variables in catalog/operator.yml
- Append the bundle render to the catalog/operator.yml
- Build the opm image which serves generated catalog

This is not pure forklift-operator migration this patch also contains
changes from https://github.com/konveyor/forklift-operator/pull/250.
The 250 PR was added for the split configuration instead of one large
CSV with multiple scripts.