Enable non-admin users to query providers from inventory
Previously, querying '/providers' by users that don't have cluster-admin
permissions failed because the request doesn't specify a UID of a
provider to query and thus, we ended up checking if the user has
permissions on all resources.

With this change, we'd always query 'providers.forklift.konveyor.io' for
'get' when a provider is specified and for 'list' otherwise. That way,
all users with ClusterRole on providers.forklift.konveyor.io can access
the data of providers from the inventory.

Signed-off-by: Arik Hadas <ahadas@redhat.com>