build: Build separate images for cold and warm migration
Later we plan to use CentoOS Stream 9 as a base image for our
containers. However, virt-v2v in C9S does not have the features needed
for warm migration. Therefore we well need to have one image for warm
migration with C8S and another image for cold migration with C9S.

For now both images will be based on C8S. The switch to C9S for cold
migration image will be done later when the controller and the container
entrypoint are prepared for the change.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>