vsphere: Allow insecure connection to the server
This is a preparation for future changes. Virt-v2v and libvirt cannot
handle SSL certificates that it cannot verify (e.g. self-signed) and
need a flag that tell it to disable checks. This does not mean we don't
need the thumbprint (certificate fingerprint). Thumbprint is still
needed by controller and will remain necessary for conversion pod when
VDDK is used.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>