Fix the step and phasing for initalization
When changing the phasing of the disk transfer before the VM creation,
the scheduling of the VM removed. Doing so, we didn't end up with the
Initialization phase, and signal the step as completed. It caused the
the phase to keep running and harmed the time counter for the upcoming
phases.

This patch ends the step and fix the issue above.

Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>