Run vSphere test in a generated namespace
Our testing framework creates namespaces to run the tests on with a
label pod-security.kubernetes.io/enforce=restricted. This makes sense
but currently we can't start the conversion pod in such namespaces and
therefore the vSphere test fails when we create the plan in a generated
namespace.

This PR removes the pod-security.kubernetes.io label from namespaces
that are generated by the testing framework and creates the resources
for the vSphere test in a generated namespace. The label should be
restored when https://github.com/kubev2v/forklift/issues/173 is fixed.

Signed-off-by: Arik Hadas <ahadas@redhat.com>