Fix Makefile when using ccache
When "ccache" is installed the PATH variable contains
"/usr/lib64/ccache" directory and the CCACHE_DIR
(default $HOME/.ccache) is used by the compiler.

However CCACHE_DIR is not writable within the bazel sandbox
and the build fails.

The proposed changes try to fix this and make sure
the CCACHE_DIR is created if it doesn't exist before
starting the build (that would make the build to fail).

Signed-off-by: Miguel Martín <mmartinv@redhat.com>