Set ownership and labels to the populator pod
This patch sets the PVC as the owner of the populator pod and the. It
also labels the OvirtVolumePopulator and the OpenstackVolumePopulator
CRs with the VM ID and the migration ID. This is useful when we will
wish to get the logs using must-gather or lookup for the relevant
populator resources.

The populator pod won't be deleted at the end of migration in order to
let us gather the logs.

Once cleanup is triggered by the controller, the CRs and populator pod
will be deleted.

When the controller starts, it will lookup based on the plan for
OpenstackVolumePopulator and OvirtVolumePopulator CRs, updating the
labels with the VM ID and migration ID. That will be handy when
upgrading to newer version while having old resources, letting them to
be cleaned up.

Owner references added to the Populator CRs and the populator pods only
after the VM is created. In order to have additional clean-up support
mostly on remote clusters when the controller won't have the ability to
do any operation after the migration.

Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>