Fix import of multi-volume VMs from vSphere
We previously named all volumes within the pod we start for binding the
PVCs with the same name, "init". Now, each volume is named after the PVC
it is expected to bind. The container is now named "main" rather than
"init" - init is a bit misleading as it might suggest we refer to an
init-container.

Signed-off-by: Arik Hadas <ahadas@redhat.com>