Attach the right PVC on different flows
When running an import with CDI, the PVC is not named based on the
source input. In that case we need to use the generated name of the PVC
to be claimed by the VM. This relevant for non volume populator imports
such as warm imports and imports to remote clusters.

Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>