virt-v2v: Build our own fixed appliance (Fixes #226)
The appliance from Kubevirt [1] is missing some packages we need. For
legal reasons it is not possible to include the missing packages there.
We need to build our own fixed appliance.

We imitate multi-stage Dockerfile build in Bazel. Unfortunately
container_run_and_extract() does not work with Podman [2] and Bazel's
linux-sandbox. Slightly less secure but working processwrapper sandbox
is used as a workaround.

[1] https://github.com/kubevirt/libguestfs-appliance/pull/17
[2] https://github.com/bazelbuild/rules_docker/issues/2251

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>