openstack: support new authentication types
Support the application credentials and token authentication types.

To choose one of the new authentication types a new field `authType`
is expected in the secret (defaults to `password` if not specified),
which can have one of the supported authentication types:

- password
- token
- v3password
- v3token
- v3applicationcredential

When choosing the application credentials authentication type
(`v3applicationcredentials`) three new secret fields are available:

- applicationCredentialID
- applicationCredentialName
- applicationCredentialSecret

One of `applicationCredentialID` or `applicationCredentialName`
is required alongside `applicationCredentialSecret`.

If `applicationCredentialName` is used it's also mandatory to
specify the `username` and `userDomainName` fields.

The application credentials must have a role assigned that permits to
read the required information (VMs, Flavors, Images, Volumes,
Snapshots, Networks...) and to stop and start VMs, (`reader` and
`member` are the typical ones). The authentication credentials
might be generated with the following command:

```
openstack application credential create \
          --role reader \
          --role member \
          --secret redhat forklift
```

When choosing the token authentication type (`token` or `v3token`)
the `projectID` or `projectName` and the `userID` or `username` are
required alongside the `token` field.

When using `projectName` it's also required to specify the `domainName`
field.

It's possible to generate the required token field with the following
command:

```
openstack token issue
```

Signed-off-by: Miguel Martín <mmartinv@redhat.com>