openstack: make sure VMs are synchronized
Attaching/detaching volumes/interfaces to a VM does not update
the `Updated` field of the VM. The consequence is the VM not
being updated within the inventory database.

The proposed patch make sure the VM is updated if any field
is changed instead of relying on the `Updated` field.

Fixes #302

Signed-off-by: Miguel Martín <mmartinv@redhat.com>