webhooks: remove superfluous error message from log line
Logging facility expect pairs of key-value arguments. Trying to log just
the error message results in a panic log message from the facility.
We could insert a key argument (e.g. "message") before the message to
fix this. But since the Error() method already logs the message for us
trying to do this on our own is superfluous.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>