controller: vmware: use ESX transport for EL9 image when Host resource exists
When there is a Host resource for the ESXi host on which the converted
VM is located we use the `esx://` transport when running virt-v2v. The
Host resource is normally created when migration network is configured
in UI for the provider. There is not switch or option whether to use
`vpx://` or `esx://` and the choice is based solely on the existence of
the Host resource. This mimics the behavior that exists with CDI.

This also fixes a bug with EL9 image. When user configured migration
network, e.g. to be used for warm migrations with CDI, this prevented
any cold migrations because we passed wrong credentials in conversion
pod secret and no thumbprint.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>