Drop unused Thumbprint field from HostSpec
This change theoretically breaks backward compatibility but it's really
unlikely that someone automated the setting of the default network of
ESX hosts, and as this field was not used (we take it from the inventory
instead), there's no change in behavior. Scripts that referred to this
field, for some reason, would need to change.

Signed-off-by: Arik Hadas <ahadas@redhat.com>