oVirt provider: Fix the POST auth request
The oVirt auth request needs to be Content-type
application/x-www-form-urlencoded, which requires the URL encoding
inside the POST data request.

For this was necessary to update the Post method in Client.
If the posted body is type string, it will pass it as it
is. If the body is any other type, it uses the json.Marshal for
json formatting, which the oVirt auth API does not support.

Signed-off-by: Martin Necas <mnecas@redhat.com>