openstack: fix project sync for new authentication methods
The project list will retrieve user projects only as
users usually lack permissions for the 'list_projects'
operation in nova.

Currently, we only allow migrations from the project
associated with the session/token/applicationcredential
so it makes no sense to sync other projects

The current implementation takes the project name from
the secret. This property maybe missing if token or
applicationcredentials authentication is used so we
need to figure out the project depending on the
authentication type and which fields are required
in each case.

Signed-off-by: Miguel Martín <mmartinv@redhat.com>