ovirt: Do not wait for timeout when snapshot removal fails
When we trigger remove-snapshot and the operation fails, we immediately
break rather than keep waiting until the timeout of waiting for the
result of the operation is reached.

This change also includes refactoring - changing the format of the error
that is generated from ConflictError and simplify the function
Client#CheckSnapshotReady.

Signed-off-by: Arik Hadas <ahadas@redhat.com>