Fix virt-v2v entrypoint script
Since we added OVA we had two options to execute the virt-v2v command.
Now the command will be built and will be executed at the end without
conditions regarding the source. This will allow us to know if we had an
error and fail the pod in such case.

Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>