Add ovirt system version to validate
This patch will check the source provider version when it's oVirt
provider and we have a direct LUN/FC to migrate. It will make sure we
can support direct LUN/FC migration with enough API details from ovirt
to do so. The minimal required version of ovirt is ovirt-engine-4.5.2.1
based on the PR:
https://github.com/oVirt/ovirt-engine/commit/e7c1f585863a332bcecfc8c3d909c9a3a56eb922

Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>