fix #437: Post VMs with desired 'Running' state
Long time ago, we posted VMs at the beginning of the migration and then
started and stopped them during the migration and therefore in order to
restore their original power state, we had to initiate another call to
patch the VM with its desired power state.

However, now that we post the VMs at the very last steps of the
migration, we can set their desired power state in their specification
and avoid another call to the target environment.

Signed-off-by: Arik Hadas <ahadas@redhat.com>