Keep the populator pod on failure
This patch will keep the populator pod existence when failing. It won't
restart or recreated. This is in the mind of aligning with v2v pod
failure and the thought that once we fail, we will most likely keep
failing. This change will allow to get the populator logs in case of
failure in a easy way.

Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>