Create TPM device based on ovirt source
When having Windows 2k22 or Windows 11 OS, a TPM device is a must. Therefore,
we will create it also for the destination VM. The data is persist in
ovirt. If supported by kubevirt the TPM will be persistent TPM, otherwise
the TPM is without persistent data.

In order to have the TPM persistent it is required to have kubvirt
v1.0.0 and adding configuration manually as described in:
https://kubevirt.io/user-guide/virtual_machines/persistent_tpm_and_uefi_state/

It is also supported in CNV running on OCP 4.14, and you need to follow:
https://access.redhat.com/documentation/en-us/openshift_container_platform/4.14/html/virtualization/virtual-machines#virt-using-vtpm-devices

Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>