Add LUKS key to the plan
This patch adds the LUKS keys to the plan specs using a secret ref. It
expected to have list of strings. Each value of the string should
contain the ID or name of the device, and the selector as pointed out in
virt-v2v docs (https://www.libguestfs.org/virt-v2v.1.html).

The secret should be provided in the destination namespace, the key
should be `luks` and the values should be the arguments provided to
virt-v2v, such as:

```
UUID:key:passphrase
UUID:clevis
```

Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>