Set transfer network on the conversion pod
Before version 2.4, the virt-v2v conversion pod operated on a local disk
after the disk was retrieved by CDI. CDI considered the transfer network
when transfering the data from vSphere.

As from version 2.4, when running a cold-migration to the local cluster,
virt-v2v is also responsible to connect to vSphere and retrieve the
disk(s) from there, therefore it now needs to consider the transfer
network that is selected as well.

This patch adds the transfer network, if set, on the conversion pod in
both cases.

Signed-off-by: Arik Hadas <ahadas@redhat.com>