Set reclaim policy of PV of OVA provider to Delete
The PV is currently not set with a reclaim policy and therefore it
defaults to 'Retain' which is conceptually incorrect since it is
supposed to be removed when the OVA provider is removed.

Here we set the reclaim policy of the PV to 'Delete' instead. While the
PV would still remain after the OVA provider is removed (since the PV is
not set with a storage class and therefore there is no driver that can
remove it), the state of the PV would change to 'Failed' rather than
'Released' in this case, which better reflects its true state. Removing
the PV for an OVA provider when the latter is removed will be addressed
separately.

Signed-off-by: Arik Hadas <ahadas@redhat.com>