Fix calls to liberr.Wrap with even number of args
As the signature of the function is:
Wrap(err error, kvpair ...interface{})
It expects to get an error and key-value pairs. However, in some places
we also pass a message after the error, which breaks this expectation.
In this PR those messages are dropped to remain with an odd number of
arguments.

Signed-off-by: Arik Hadas <ahadas@redhat.com>