Ensure namespace when validating VDDK
In order to validate the VDDK init image, we create a job on the
destination provider. However, that would fail when the target namespace
doesn't exist - it would be created only when the plan is triggered. So
now we ensure the target namespace exists also during the validation of
the VDDK init image.

Signed-off-by: Arik Hadas <ahadas@redhat.com>