Fix validation of plans with an invalid source provider
When validating a plan we have specific logic that applies for source
providers of type OpenShift but the check of the type of the source
provider didn't take into consideration that the source provider may be
invalid, leading to a sig fault. Here we extend the function that
validates the source providers of type OpenShift with nil-check of the
source provider as done in other validation function for a plan.

Signed-off-by: Arik Hadas <ahadas@redhat.com>