vSphere: Fix security context of VDDK validation pod
The previous attempt to fix the execution of the VDDK validation pod
didn't succeed - it allowed to run it on the 'default' namespace but its
execution failed on other retricted namespaces in which the QEMU user
(107) is not included in the allowed range.

In order to fix this properly on OpenShift, we drop RunAsNonRoot and
RunAsUser so that OpenShift would pick a proper user based on the
configuration of the namespace.

Signed-off-by: Arik Hadas <ahadas@redhat.com>