Add forklift disk ID annotation to DV and PVCs
In OCP 4.15, CDI dropped their annotation of the import disk ID.
We added an annotation `forklift.konveyor.io/disk-source` before to the
DataVolumes only which propagates to the PersistentVolumeClaim. Now it
will also be included on other non-CDI flows.

This change is made as we used the CDI annotation to be an identifier
when importing a disk. Not having it caused an NPE. Now we will use the
forklift annotation.

Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>