vsphere: transfer updated disk state on remote migrations
The problem these changes address is as follow: when initiating a cold
migration from vSphere to a remote provider of a VM with a snapshot(s),
the transferred disk didn't include the most updated state of the disk
because the very first volume of the disk, that doesn't include later
changes, was transferred.

Unlike warm migrations, in which we intend to transfer the first (base)
volume of the disk initially and then copy the changes using CBT, in
cold migrations we need to transfer the state of the very last volume to
include the most updated state of the disk.

In Forklift 2.3 and below, this issue affected all cold migrations. As
from 2.4, it only affects cold migrations to a remote cluster, in which
we still use CDI to transfer the disks from vSphere.

Signed-off-by: Arik Hadas <ahadas@redhat.com>