vSphere: fix crash due to disk without a datastore
Previously, we assumed all disks of types
VirtualDiskFlatVer1BackingInfo, VirtualDiskFlatVer2BackingInfo and
VirtualDiskRawDiskMappingVer1BackingInfo are located on a datastore. The
only exception was disks of type VirtualDiskRawDiskVer2BackingInfo for
which we didn't assume they are located on a datastore but for them RDM
is set to true and so we have a critical concern saying VMs which such
disks cannot be migrated.

However, the 'Datastore' property is optional and according to the
vSphere documentation: "If the file is not located on a datastore, then
this reference is null.". We noticed that this could actually happen,
even though we were not able to reproduce it on our development
environments, and once it happens forklift-controller crashes while
creating the inventory.

Therefore this PR introduces the following changes:
1. Make the code that builds the inventory more robust in terms of
   handling VMs with disks that are not located on a datastore
2. Add a critical concern for such VMs with disks that are not located
   on a datastore, saying such VMs cannot be migrated (later on we will
   anyway fail the validation of a plan due to unmapped storage in this
   case, even if we don't block migration of VMs with critical concerns)

Signed-off-by: Arik Hadas <ahadas@redhat.com>