operator: change matchLabels of operator deployment
The matcher includes the label "name: controller-manager". This differs
from what we set in MTV (there 'name' is mapped to 'forklift-operator')
and this can lead to errors when creating a new release of MTV.

Here we change the aforementioned label to match that of MTV. This
change would break upgradea to Forklift 2.7.0, however, there is an easy
workaround for this by removing the custom resource
'forklift-controller' of type ForkliftController from the namespace the
operator is deployed on.

Signed-off-by: Arik Hadas <ahadas@redhat.com>