ova: fix the version of encoding/gob
Unfortunately, we rely on 'encoding/gob' to determine the UUID of VMs
from OVAs and this package has changed in Go 1.21 in a way we get a
different UUID compared to the one we get with Go 1.20 for the same
input, which could break existing OVA providers.

Therefore, this package integrates the relevant pieces from
'encoding/gob' of Go 1.20 so that we can bump Go to 1.21 without
breaking compatibility with OVA providers that were defined in lower
versions of Forklift.

Signed-off-by: Arik Hadas <ahadas@redhat.com>