Set content-type of tls-certificate response
The content-type of the response that included the TLS certificate
wasn't set, and therefore static analysis tools could have thought we
intend to return text/html that wasn't sanitized. However, in this case
the return value is not in the form of text/html but text/plain. Now the
content-type is set accordingly to avoid that and so that clients won't
treat the output as HTML.

This commit also reverts 3848777a.

Signed-off-by: Arik Hadas <ahadas@redhat.com>