Fix virt-v2v progress update after switching to go wrapper
After switching to use a Go wrapper, not all the information
was redirected to the virt-v2v monitor command. Additionally, with changes
done in the virt-v2v progress display, the progress was not reported to the controller.

This patch fixes the piping to the virt-v2v monitoring command
and the regex used to determine the copy disk progress of the migration.

Signed-off-by: Bella Khizgiyaev <bkhizgiy@redhat.com>