vsphere: load hosts per source provider
When looking for the configuration of an ESXi host that a certain VM
resides on, we query all the hosts in the namespace and then looking for
the host by its MOR (Managed Object Reference). This almost always works
but it could be wrong when there are several providers within the same
namespace that refer to the same host. This is more likely to happen
during tests where we define two providers with different propreties for
the same vSphere environment. Therefore, adding another check that
ensures the host refers to the provider which is the source provider on
the plan of the migrated VM, otherwise skip that host.

Signed-off-by: Arik Hadas <ahadas@redhat.com>