vsphere: perform CBT operations on hosts when needed
Previously, when running warm migrations from a vSphere provider with
vCenter as its SDK endpoint, we ran create-snapshot, set-checkpoints and
remove-snapshots on vCenter, even when there was a network defined for
an ESXi host. That cause CDI to fail: the identifier of a snapshot that
was recievd from vCenter was different than the identifier of the same
snapshot in ESXi and therefore CDI, that interacts with ESXi hosts in
this scenario, failed to find the snapshot.

Now, when running CBT operations we check whether the disk(s) will be
transferred from vCenter or ESXi, and perform the aforementioned CBT
operations on the corresponding component.

Signed-off-by: Arik Hadas <ahadas@redhat.com>