Build operator with rules_docker instead of containerfile
We cannot push the operator image with recent images of Ubuntu in GitHub
actions because recent versions of Docker create the image differently
and as a result, they cannot be pushed to quay.io. Thus, replacing the
way we build the operator image to match how other images are built and
this way, we are able to push the image to quay.io.

Signed-off-by: Arik Hadas <ahadas@redhat.com>