Fix migration of OVA to WFFC storage classes
In b19098a we've added the following logic: when using storage classes
that are set with WaitForFirstConsumer in migrations from vSphere, we
treat the PendingPopulation status as Succeeded since the conversion pod
would be the first consumer of the PVCs. This PR applies the same logic
also to migrations from OVA providers.

Signed-off-by: Arik Hadas <ahadas@redhat.com>